fix bug about typescript
