+++ models/news.ts
@@ -1,0 +1,8 @@
+export interface News {
+    id: string;
+    title: string;
+    description: string;
+    photo: string;
+    openFlag: boolean;
+    delFlag: boolean;
+}
